+++ src/Middleware/StratigilityMiddleware.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rougin\Slytherin\Middleware;
+
+use Zend\Stratigility\MiddlewarePipe;
+
+/**
+ * Stratigility Middleware
+ *
+ * A simple implementation of middleware that is built on top of Zend's Stratigility.
+ * 
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class StratigilityMiddleware extends MiddlewarePipe implements MiddlewareInterface {}